refactor: extract sanitizeString into @monkeytype/util/strings
