+++ backend/__tests__/utils/misc.spec.ts
@@ -197,43 +197,6 @@ describe("Misc Utils", () => {
     );
   });
 
-  it("sanitizeString", () => {
-    const testCases = [
-      {
-        input: "h̶̼͔̭͈̏́̀́͋͜ͅe̵̺̞̦̫̫͔̋́̅̅̃̀͝͝ļ̶̬̯͚͇̺͍̞̫̟͖͋̓͛̆̒̓͜ĺ̴̗̘͇̬̆͂͌̈͊͝͝ỡ̴̡̦̩̠̞̐̃͆̚͠͝",
-        expected: "hello",
-      },
-      {
-        input: "hello",
-        expected: "hello",
-      },
-      {
-        input: "hel   lo",
-        expected: "hel  lo",
-      },
-      {
-        input: "   hel   lo   ",
-        expected: "hel  lo",
-      },
-      {
-        input: "",
-        expected: "",
-      },
-      {
-        input: "   \n\n\n",
-        expected: "",
-      },
-      {
-        input: undefined,
-        expected: undefined,
-      },
-    ];
-
-    testCases.forEach(({ input, expected }) => {
-      expect(Misc.sanitizeString(input)).toEqual(expected);
-    });
-  });
-
   it("getOrdinalNumberString", () => {
     const testCases = [
       {
